Added proper serial recognition
Turns out that the serial.Serial.isOpen() method did not register
that the serial port was not open because of the USB adapter.
Since the usb adapter is still connected the actual file descriptor
is open. So isOpen always returns true even if the actual serial
cable is not true.

To fix this we check the status of the DSR pin on the RS232 connector.
By the RS232 standard the DSR pin is always high when devices are
connected to indicate device connectivity.

Additionally, the serial port is opened each time we send a print
job rather than continuously.