Updated setup script
Because of new deployment techniques the setup scripts has been changed
to be executed remotely. A relative paths were fixed and the execution
semantics changed. The script now takes an arbitrary number of printer
IDs as arguments instead of being an interactive prompt. This is for
ease of execution of the remote server.