Added retry to variable set
When setting variables on the modem we get the variable
from the modem until the gotten value and the attempted
set are equal. This is because the modem CLI is hot garbage
and always returns 255, so this is the only way to really
tell when a value is set properly.